Latest 3.1.x Merge (#3)
* Drop the timer in Ultimately for an Executor

Change the behaviour of Ultimately to spawn a single daemon thread for
handling dispatching of futures.

* Move ulimately scheduler to object

* Added a type annotation for clarity in Ultimately and added a tolerance to an UltimatelySpec test to reduce the likelihood of flicker.

* Get NonEmptyArray and its tests to compile in scala 2.10.

* Removed the use of deprecated getStackTraceString in FuturesSpec.

* Enhanced README.md to include instruction to build and publish for Scala 2.13.

* Changed default scala version to 2.11.11.

* Fixed occasionally failing tests in PosIntSpec and PosZIntSpec when tests run under scala 2.10, which is caused by range size cannot be more than Int.MaxValue, added code to make sure the limit is not violated in the test.

* Fixed stack depth problem in RefSpecLike and fixture.SpecLike under Scala 2.12 and 2.13.

* Removed special treatment of validating method for scala 2.12.

* Extends special handle of Predef string printing to scala 2.13 in RequirementsSpec.scala.

* Extends special handle of Predef string printing to scala 2.13 in AssertionsSpec.scala.

* Make source compatible to build with Scala 2.10, 2.11, 2.12 and 2.13.0-M2.

Conflicts:

	project/build.properties
	project/plugins.sbt
	project/scalatest.scala

* Fixed failing tests in NegIntSpec, NegZIntSpec and NonZeroIntSpec when built under Scala 2.10.

* Fixed failing tests in PosIntSpec and PosZIntSpec when built under Scala 2.10.

* Make examples.js runnable by not copying examples that use java.io.File when generating examples.js.

* Make source compatible to build with Scala 2.10, 2.11, 2.12 and 2.13.0-M2.

* In travis, bumped up scala versions, removed jdk7 build, added add-on to install jdk6, make builds for scala 2.10 and 2.11 to use jdk6 and 2.12 to use jdk8.

* Added dist: trusty and change sudo value to required.

* Got the right set of external dependencies version to get genRegularTests5 passes.

* Got genRegularTests5 to work again.

* Increased stack size to 10M in travis_build.sh.

* Changed JVM_OPTS to SBT_OPTS in travis_build.sh.

* Added scala 2.13.0-M2 to travis build.

* Fixed genInspectorsShorthandsTests2 problem when runs under scala 2.13.0-M2.

* Added scala version to gentests shared settings.

* Make sure compile in travis is using 10M stack size.

* Rewritten time limited examples to be scala-js friendly.

* Changed back sudo to false in .travis.yml, trying to fix some test in the matrix failed before it started problem.

* Changed to use scala 2.12.4 as default scala version.

* Changed getJavaHome in build to throw error when Java 6 is not used to build for Scala 2.10 and 2.11.

* Updated .travis.yml and README.md to use scala 2.12.4.

* Commentted out SbtCommandParser, which we currently doesn't support, to get rid of combinator parser dependency that it brought in.

* Should exclude scala 2.12.4 not 2.12.3 in travis.

* Made scalactic and scalatest project to generate scala source directly under sourceManaged root, making them friendlier to IDE.

* Work in progress fixing sourceManaged problem for GenRegularTests.

* Fixed sourceManaged up to must matchers tests generators.

* Fixed sourceManaged problem for all generated tests.

* Completed fixing remaining projects for sourceManaged.

* Exclude Ultimately from scalatest-js.

* Adjusted command to run generated tests in README.md.
(cherry picked from commit 3803aaeac91e4f8313ec66600de24637dcb31ac5)

* Make generators to generate files only when necessary.
(cherry picked from commit 68d9c297f6da885dc85a27c8c55e3274c01e860a)

Conflicts:

	project/GenTable.scala

* Made all test generators and scala-js generators to generate files only when necessary, completed this behavior now.
(cherry picked from commit a996db8eb07046efac2a8068ea0dfa536e049272)

* Fixed Tolerance related failing tests.

+++ project/GenRegularTestBase.scala
@@ -24,22 +24,26 @@ trait GenRegularTestsBase {
   val baseDirPath: String
   val name: String
 
-  def copyFile(inputFile: File, outputFile: File) {
-    val writer = new BufferedWriter(new FileWriter(outputFile))
-    try {
-      val inputLines = Source.fromFile(inputFile).getLines().toList // for 2.8
-      for (line <- inputLines) {
-        writer.write(line.toString)
-        writer.newLine() // add for 2.8
+  def copyFile(inputFile: File, outputFile: File): File = {
+    if (!outputFile.exists || inputFile.lastModified > outputFile.lastModified) {
+      val writer = new BufferedWriter(new FileWriter(outputFile))
+      try {
+        val inputLines = Source.fromFile(inputFile).getLines().toList // for 2.8
+        for (line <- inputLines) {
+          writer.write(line.toString)
+          writer.newLine() // add for 2.8
+        }
+      }
+      finally {
+        writer.flush()
+        writer.close()
+        println("Generated " + outputFile.getAbsolutePath)
       }
     }
-    finally {
-      writer.close()
-      println("Generated " + outputFile.getAbsolutePath)
-    }
+    outputFile
   }
 
-  def copyFile(targetBaseDir: File, filePath: String): Unit = {
+  def copyFile(targetBaseDir: File, filePath: String): File = {
     val source = new File(filePath)
     val targetDir = new File(targetBaseDir, source.getPath.substring("scalatest-test/src/test/scala/".length, source.getPath.lastIndexOf("/")))
     targetDir.mkdirs()
@@ -47,41 +51,50 @@ trait GenRegularTestsBase {
     copyFile(source, target)
   }
 
-  def processDir(dir: File, targetDir: File): Unit = {
+  def processDir(dir: File, targetDir: File): Seq[File] = {
     targetDir.mkdirs()
-    for (sourceFile <- dir.listFiles) {
+    val (sourceFiles, subDirs) = dir.listFiles.partition(_.isFile)
+    sourceFiles.flatMap { sourceFile =>
       val sourceFileName = sourceFile.getName
-      if (includeFile(sourceFile)) {
+      if (sourceFileName.endsWith(".scala") && includeFile(sourceFile)) {
         val outputFile = new File(targetDir, sourceFileName)
-        copyFile(new File(dir, sourceFileName), outputFile)
+        Seq(copyFile(new File(dir, sourceFileName), outputFile))
       }
-      else if (sourceFile.isDirectory)
-        processDir(sourceFile, new File(targetDir, sourceFileName))
+      else
+        Seq.empty[File]
+    } ++
+    subDirs.flatMap { subDir =>
+      processDir(subDir, new File(targetDir, subDir.getName))
     }
   }
 
-  def processJavaDir(dir: File, targetDir: File): Unit = {
+  def processJavaDir(dir: File, targetDir: File): Seq[File] = {
     targetDir.mkdirs()
-    for (sourceFile <- dir.listFiles) {
+    val (sourceFiles, subDirs) = dir.listFiles.partition(_.isFile)
+    sourceFiles.flatMap { sourceFile =>
       val sourceFileName = sourceFile.getName
-      if (sourceFile.isFile) {
+      if (sourceFileName.endsWith(".java")) {
+        val sourceFile = new File(dir, sourceFileName)
         val outputFile = new File(targetDir, sourceFileName)
-        copyFile(new File(dir, sourceFileName), outputFile)
+        if (!outputFile.exists || sourceFile.lastModified > outputFile.lastModified)
+          copyFile(sourceFile, outputFile)
+        Seq(outputFile)
       }
-      else if (sourceFile.isDirectory)
-        processJavaDir(sourceFile, new File(targetDir, sourceFileName))
+      else
+        Seq.empty[File]
+    } ++
+    subDirs.flatMap { subDir =>
+      processJavaDir(subDir, new File(targetDir, subDir.getName))
     }
   }
 
-  def copyJavaDir(targetBaseDir: File): Unit = {
-    val testDir = targetBaseDir.getParentFile.getParentFile
-    val javaDir = new File(testDir, "java")
-    processJavaDir(new File("scalatest-test/src/test/java"), new File(javaDir, name))
+  def copyJavaDir(targetBaseDir: File): Seq[File] = {
+    processJavaDir(new File("scalatest-test/src/test/java"), targetBaseDir)
   }
 
-  def genTest(targetBaseDir: File, version: String, scalaVersion: String) {
+  def genTest(targetBaseDir: File, version: String, scalaVersion: String): Seq[File] = {
     val sourceBaseDir = new File(baseDirPath)
-    val regularDir = new File(targetBaseDir, name + "/org/scalatest/")
+    val regularDir = new File(targetBaseDir, "/org/scalatest/")
 
     processDir(sourceBaseDir, regularDir)
   }
